CHANGE: Put results into a separate class
This will make it easier to pass results with a lot of information (f.e.
also including baselines and std) in the future. It also cleans up the
code a bit, as plotting and saving results are now delegated to the
results classes and don't clutter up the experiment running code.