CHANGE: Simplified hierarchy handling
The Hierarchy list has now been moved to a tree-like structure, which
replaces the previous HierarchyTree and HierarchyList. Simplifies
handling and calculation of AARI, as well as enabling us to now work
with unbalanced hierarchies. This renders the notebook for exp20
unusable, since quite some code for the old method of handling
hierarchies has been removed.